+++ plugins/helper.go
@@ -34,3 +34,12 @@ func getYouTubeCookie(ctx *Context) string {
 	cookie, _ := s.GetSetting(ctx.Ctx, "youtube_cookie")
 	return cookie
 }
+
+func isWordPrefix(s string) bool {
+	for _, r := range s {
+		if (r >= 'a' && r <= 'z') || (r >= 'A' && r <= 'Z') || (r >= '0' && r <= '9') {
+			return true
+		}
+	}
+	return false
+}